Migrate recommender to backend
The recommender now relies on the backend, so this adds an API call to do so
Removed all useless files from the frontend
